requirements.txt
feedparser>=6.0.10
openai>=1.6.1
PyYAML>=6.0.1
requests>=2.31.0
python-dateutil>=2.8.2
